Added new client implementation build on useSyncExternalStore
