+++ client-react/package.json
@@ -23,6 +23,7 @@
   },
   "dependencies": {
     "@krmx/base": "^0.5.4",
+    "@krmx/client": "^0.5.0",
     "immer": "^10.0.3"
   },
   "peerDependencies": {
